+++ tests/tsp/plot.py
